+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Updated `nalgebra` to 0.10.
 * Added `Solve` trait for solving a linear matrix equation.
 * Added `Inverse` for computing the multiplicative inverse of a matrix.
+* Added `Cholesky` for decomposing a positive-definite matrix.